1번 회원 관리자로 지정, 수정 삭제시 권한체크

+++ src/main/java/com/nftf/kys/contoller/UsrArticleController.java
@@ -81,7 +81,13 @@ public class UsrArticleController {
 
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(Integer id) {
+	public ResultData doDelete(Integer id, HttpSession session) {
+		int loginedMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);		
+		
+		if( loginedMemberId == 0) {
+			return new ResultData("F-2", "로그인 후 이용해주세요.");
+		}		
+		
 		if (id == null) {
 			return new ResultData("F-1", "id를 입력해주세요.");
 		}
@@ -91,13 +97,25 @@ public class UsrArticleController {
 		if (article == null) {
 			return new ResultData("F-1", "해당 게시물은 존재하지 않습니다.");
 		}
+		
+		ResultData actorCanModifyRd = articleService.getActorCanDelete(article, loginedMemberId);
+		
+		if ( actorCanModifyRd.isFail()) {
+			return actorCanModifyRd;			
+		}
 
 		return articleService.deleteArticle(id);
 	}
 
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData doModify(Integer id, String title, String body) {
+	public ResultData doModify(Integer id, String title, String body, HttpSession session) {		
+		int loginedMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);		
+		
+		if( loginedMemberId == 0) {
+			return new ResultData("F-2", "로그인 후 이용해주세요.");
+		}
+		
 		if (id == null) {
 			return new ResultData("F-1", "id를 입력해주세요.");
 		}
@@ -114,6 +132,12 @@ public class UsrArticleController {
 
 		if (article == null) {
 			return new ResultData("F-1", "해당 게시물은 존재하지 않습니다.");
+		}
+		
+		ResultData actorCanModifyRd = articleService.getActorCanModify(article, loginedMemberId);
+		
+		if ( actorCanModifyRd.isFail()) {
+			return actorCanModifyRd;			
 		}
 
 		return articleService.modifyArticle(id, title, body);
